Remove support for nonorientable PantsDecompositions
Reasons: For orientable surfaces, the number of inner pants curves stays
constant, so one can assume that the inner pants curves are numbered
from 1 to K. This is very convenient, since the switches of
Dehn-Thurston train tracks can just get the same number as for
the corresponding inner pants curve. Not only the coding is easier, but
the code runs faster.

There are also a lot of much more important features to implement first,
before nonorientable surfaces are supported. (If they will ever get
supported. The elementary moves and Dehn-Thurston train tracks
become fairly technical to work out, so maybe the demand won't be
high enough to justify the work put it.)